Foi alterado o nome de algumas rotas para a criação da rota pricelistId e também ajustado o tipo de algumas variaveis que estavam incompativeis na hora de receber no flutter
