+++ back-end-leilaorv/src/controllers/getProductPriceList.ts
@@ -1,19 +1,16 @@
 import { Router } from 'express';
 import { AppDataSource } from '../config/db.datasource';
-import { Product } from '../models/product';
-import { Enterprise } from '../models/enterprise';
 import { PriceList } from '../models/price_list';
 import { Request, Response } from 'express';
-import { authenticateToken } from '../middleware/authenticateToken';
 
-const getEnterpriseProductsByProductId = async (req: Request, res: Response): Promise<void> => {
+const getProductPriceList = async (req: Request, res: Response): Promise<void> => {
     const { productId } = req.params;
 
     try {
-        const enterpriseProductsRepository = AppDataSource.getRepository(PriceList);
+        const priceListReposity = AppDataSource.getRepository(PriceList);
         
         // Como 'product' é um relacionamento, precisamos passar um objeto para o campo 'where'
-        const enterpriseProducts = await enterpriseProductsRepository.find({
+        const productInlist = await priceListReposity.find({
             where: {
                 product: {
                     id: parseInt(productId, 10), // Certifique-se de converter para número
@@ -22,22 +19,28 @@ const getEnterpriseProductsByProductId = async (req: Request, res: Response): Pr
             relations: ['enterprise', 'product'],
         });
 
-        if (!enterpriseProducts || enterpriseProducts.length === 0) {
+        if (!productInlist || productInlist.length === 0) {
             res.status(404).json({ message: 'Nenhuma associação encontrada para o produto' });
             return;
         }
 
         // Formatando o resultado para retornar somente as informações necessárias
-        const formattedEnterpriseProducts = enterpriseProducts.map(ep => ({
+        const formatedPriceList = productInlist.map(ep => ({
             //nome: ep.product?.nomeProd,
           //  tamanho: ep.product?.tamanhoProd,
             //marca: 'Marca Genérica', // Placeholder, caso precise de uma informação que não existe no modelo
-            user: ep.user?.nome,
-            empresa: ep.enterprise?.name,
-            preco: ep.price,
+            nameProd: ep.product.name,
+            mark:ep.product.mark,
+            price: ep.price,     
+            isSale: ep.isSale,
+            enterprise: ep.enterprise?.name,
+            un:ep.product.unidade_measure,
+            weigth:ep.product.weigth,
+            img_url:ep.product.img_url        
+           
         }));
 
-        res.status(200).json(formattedEnterpriseProducts);
+        res.status(200).json(formatedPriceList);
     } catch (err) {
         console.error(err);
         res.status(500).json({ message: 'Erro ao buscar dados da associação enterprise_products' });
@@ -45,4 +48,4 @@ const getEnterpriseProductsByProductId = async (req: Request, res: Response): Pr
 };
 
 // Exportando o método para ser usado nas rotas
-export { getEnterpriseProductsByProductId };
+export { getProductPriceList };
